HTML TAGS
HTML Tags
If you want to build a beautiful website, tags are essential elements that help you achieve that.
An HTML tag acts as a container for content or other HTML tags. Tags are words enclosed within < and > angle brackets.
They serve as keywords that instruct the web browser on how to format and display the content.

Paired and Unpaired HTML Tags
Well, that was a really long list. Don't worry we will study these in detail. In HTML, tags can be broadly categorized into two types:

1. Paired Tags (Container Tags)
These are tags that come in pairs, consisting of an opening tag and a corresponding closing tag. The content goes between these two tags.

Opening Tag: The opening tag starts with < and ends with >. For example, <p>.
Closing Tag: The closing tag also starts with < but includes a forward slash / before the tag name, and ends with >. For example, </p>.

Examples:
Paragraphs: <p>This is a paragraph.</p>
Headings: <h1>This is a heading.</h1> 
2. Unpaired Tags (Self-Closing Tags or Stand-Alone Tags)
These are tags that don't require a closing tag. They are self-contained, encapsulating all the information within a single tag.

Self-Closing Tag: A self-closing tag starts with < and ends with /> (though the / is optional in HTML5). For example, <img /> or <br>.
Note: Later if you happen to use react or a framework like Next.js, you will have to close the tag like this <br/> <hr/>. So it is better to cultivate the habit!

Examples of self-closing tags:
Line Break: <br/>
Horizontal Rule: <hr/>
Image: <img src="image.jpg" alt="An example image"/>